+++ mex/backend/merged/helpers.py
@@ -1,23 +1,24 @@
-from typing import Annotated, Any, cast
+from typing import Annotated, Any, Literal, overload
 
-from pydantic import Field, TypeAdapter, ValidationError
+from pydantic import Field, TypeAdapter
+from pydantic_core import ValidationError
 
-from mex.backend.constants import NUMBER_OF_RULE_TYPES
-from mex.backend.fields import MERGEABLE_FIELDS_BY_CLASS_NAME
 from mex.backend.graph.connector import GraphConnector
-from mex.backend.graph.exceptions import InconsistentGraphError
-from mex.backend.merged.models import MergedItemSearch
+from mex.backend.merged.exceptions import MergingError
+from mex.backend.merged.models import MergedItemSearch, PreviewItemSearch
 from mex.backend.rules.helpers import transform_raw_rules_to_rule_set_response
-from mex.backend.utils import extend_list_in_dict, prune_list_in_dict, reraising
-from mex.common.exceptions import MExError
+from mex.backend.utils import extend_list_in_dict, prune_list_in_dict
+from mex.common.fields import MERGEABLE_FIELDS_BY_CLASS_NAME
 from mex.common.models import (
     EXTRACTED_MODEL_CLASSES_BY_NAME,
     MERGED_MODEL_CLASSES_BY_NAME,
+    PREVIEW_MODEL_CLASSES_BY_NAME,
     RULE_MODEL_CLASSES_BY_NAME,
     AnyAdditiveModel,
     AnyExtractedModel,
     AnyMergedModel,
     AnyPreventiveModel,
+    AnyPreviewModel,
     AnyRuleSetRequest,
     AnyRuleSetResponse,
     AnySubtractiveModel,
@@ -91,33 +92,65 @@ def _apply_subtractive_rule(
         prune_list_in_dict(merged_dict, field_name, rule_value)
 
 
+@overload
 def create_merged_item(
     identifier: Identifier,
     extracted_items: list[AnyExtractedModel],
     rule_set: AnyRuleSetRequest | AnyRuleSetResponse | None,
-) -> AnyMergedModel:
+    validate_cardinality: Literal[False],
+) -> AnyPreviewModel: ...
+
+
+@overload
+def create_merged_item(
+    identifier: Identifier,
+    extracted_items: list[AnyExtractedModel],
+    rule_set: AnyRuleSetRequest | AnyRuleSetResponse | None,
+    validate_cardinality: Literal[True],
+) -> AnyMergedModel: ...
+
+
+def create_merged_item(
+    identifier: Identifier,
+    extracted_items: list[AnyExtractedModel],
+    rule_set: AnyRuleSetRequest | AnyRuleSetResponse | None,
+    validate_cardinality: Literal[True, False],
+) -> AnyPreviewModel | AnyMergedModel:
     """Merge a list of extracted items with a set of rules.
 
     Args:
         identifier: Identifier the finished merged item should have
         extracted_items: List of extracted items, can be empty
         rule_set: Rule set, with potentially empty rules
+        validate_cardinality: Merged items validate the existence of required fields and
+            the lengths of lists, set this to `False` to avoid this and return a
+            "preview" of a merged item instead of a valid merged item
 
     Raises:
-        InconsistentGraphError: When the graph response cannot be parsed
+        MergingError: When the given items cannot be merged
 
     Returns:
-        Instance of a merged item
+        Instance of a merged or preview item
     """
+    model_class_lookup: (
+        dict[str, type[AnyPreviewModel]] | dict[str, type[AnyMergedModel]]
+    )
+    if validate_cardinality:
+        model_prefix = "Merged"
+        model_class_lookup = MERGED_MODEL_CLASSES_BY_NAME
+    else:
+        model_prefix = "Preview"
+        model_class_lookup = PREVIEW_MODEL_CLASSES_BY_NAME
+
     if rule_set:
-        entity_type = ensure_prefix(rule_set.stemType, "Merged")
+        entity_type = ensure_prefix(rule_set.stemType, model_prefix)
     elif extracted_items:
-        entity_type = ensure_prefix(extracted_items[0].stemType, "Merged")
+        entity_type = ensure_prefix(extracted_items[0].stemType, model_prefix)
     else:
         msg = "One of rule_set or extracted_items is required."
-        raise MExError(msg)
+        raise MergingError(msg)
     fields = MERGEABLE_FIELDS_BY_CLASS_NAME[entity_type]
-    cls = MERGED_MODEL_CLASSES_BY_NAME[entity_type]
+    cls = model_class_lookup[entity_type]
 
     merged_dict: dict[str, Any] = {"identifier": identifier}
 
@@ -127,59 +160,97 @@ def create_merged_item(
     if rule_set:
         _apply_additive_rule(merged_dict, fields, rule_set.additive)
         _apply_subtractive_rule(merged_dict, fields, rule_set.subtractive)
-    merged_item = reraising(
-        ValidationError,
-        InconsistentGraphError,
-        cls.model_validate,
-        merged_dict,
-    )
-    return cast(AnyMergedModel, merged_item)  # mypy, get a grip!
+    try:
+        return cls.model_validate(merged_dict)
+    except ValidationError as error:
+        raise MergingError from error
 
 
-def merge_search_result_item(item: dict[str, Any]) -> AnyMergedModel:
+@overload
+def merge_search_result_item(
+    item: dict[str, Any],
+    validate_cardinality: Literal[False],
+) -> AnyPreviewModel: ...
+
+
+@overload
+def merge_search_result_item(
+    item: dict[str, Any],
+    validate_cardinality: Literal[True],
+) -> AnyMergedModel: ...
+
+
+def merge_search_result_item(
+    item: dict[str, Any],
+    validate_cardinality: Literal[True, False],
+) -> AnyPreviewModel | AnyMergedModel:
     """Merge a single search result into a merged item.
 
     Args:
         item: Raw merged search result item from the graph response
+        validate_cardinality: Merged items validate the existence of required fields and
+            the lengths of lists by default, set this to `False` to avoid this and
+            return a "preview" of a merged item instead of a valid merged item
 
     Raises:
-        InconsistentGraphError: When the graph response item has inconsistencies
+        MergingError: When the given items cannot be merged
 
     Returns:
-        AnyMergedModel instance
+        Instance of a merged or preview item
     """
     extracted_items = [
         EXTRACTED_MODEL_ADAPTER.validate_python(component)
         for component in item["components"]
         if component["entityType"] in EXTRACTED_MODEL_CLASSES_BY_NAME
     ]
-    rules_raw = [
+    raw_rules = [
         component
         for component in item["components"]
         if component["entityType"] in RULE_MODEL_CLASSES_BY_NAME
     ]
-    if len(rules_raw) == NUMBER_OF_RULE_TYPES:
-        rule_set_response = transform_raw_rules_to_rule_set_response(rules_raw)
-    elif len(rules_raw) == 0:
-        rule_set_response = None
+    if raw_rules:
+        rule_set = transform_raw_rules_to_rule_set_response(raw_rules)
     else:
-        msg = f"Unexpected number of rules found in graph: {len(rules_raw)}"
-        raise InconsistentGraphError(msg)
+        rule_set = None
 
     return create_merged_item(
-        identifier=item["identifier"],
+        identifier=Identifier(item["identifier"]),
         extracted_items=extracted_items,
-        rule_set=rule_set_response,
+        rule_set=rule_set,
+        validate_cardinality=validate_cardinality,
     )
 
 
+@overload
 def search_merged_items_in_graph(
     query_string: str | None = None,
     stable_target_id: str | None = None,
     entity_type: list[str] | None = None,
     skip: int = 0,
     limit: int = 100,
-) -> MergedItemSearch:
+    validate_cardinality: Literal[False] = False,
+) -> PreviewItemSearch: ...
+
+
+@overload
+def search_merged_items_in_graph(
+    query_string: str | None = None,
+    stable_target_id: str | None = None,
+    entity_type: list[str] | None = None,
+    skip: int = 0,
+    limit: int = 100,
+    validate_cardinality: Literal[True] = True,
+) -> MergedItemSearch: ...
+
+
+def search_merged_items_in_graph(  # noqa: PLR0913
+    query_string: str | None = None,
+    stable_target_id: str | None = None,
+    entity_type: list[str] | None = None,
+    skip: int = 0,
+    limit: int = 100,
+    validate_cardinality: Literal[True, False] = True,
+) -> PreviewItemSearch | MergedItemSearch:
     """Search for merged items.
 
     Args:
@@ -188,12 +259,15 @@ def search_merged_items_in_graph(
         entity_type: Optional entity type filter
         skip: How many items to skip for pagination
         limit: How many items to return at most
+        validate_cardinality: Merged items validate the existence of required fields and
+            the lengths of lists by default, set this to `False` to avoid this and
+            return "previews" of merged items instead of valid merged items
 
     Raises:
-        InconsistentGraphError: When the graph response has inconsistencies
+        MergingError: When the given items cannot be merged
 
     Returns:
-        MergedItemSearch instance
+        Search response for preview or merged items
     """
     graph = GraphConnector.get()
     result = graph.fetch_merged_items(
@@ -204,19 +278,14 @@ def search_merged_items_in_graph(
         limit=limit,
     )
     total: int = result["total"]
-    items: list[AnyMergedModel] = [
-        reraising(
-            ValidationError,
-            InconsistentGraphError,
-            merge_search_result_item,
+    items = [
+        merge_search_result_item(
             item,
+            validate_cardinality=validate_cardinality,
         )
         for item in result["items"]
     ]
-    return reraising(
-        ValidationError,
-        InconsistentGraphError,
-        MergedItemSearch,
-        items=items,
-        total=total,
-    )
+
+    if validate_cardinality:
+        return MergedItemSearch(items=items, total=total)
+    return PreviewItemSearch(items=items, total=total)