+++ tests/graph/test_query.py
@@ -12,6 +12,14 @@ def query_builder() -> QueryBuilder:
         merged_labels=["MergedThis", "MergedThat", "MergedOther"],
         nested_labels=["Link", "Text", "Location"],
         rule_labels=["AdditiveThis", "AdditiveThat", "AdditiveOther"],
+        extracted_or_rule_labels=[
+            "ExtractedThis",
+            "ExtractedThat",
+            "ExtractedOther",
+            "AdditiveThis",
+            "AdditiveThat",
+            "AdditiveOther",
+        ],
     )
     return builder
 
@@ -31,7 +39,7 @@ def test_create_full_text_search_index(query_builder: QueryBuilder) -> None:
         search_fields=["texture", "sugarContent", "color"],
     )
     assert (
-        query
+        str(query)
         == """\
 CREATE FULLTEXT INDEX search_index IF NOT EXISTS
 FOR (n:Apple|Orange)
@@ -45,7 +53,7 @@ def test_create_identifier_uniqueness_constraint(query_builder: QueryBuilder) ->
         node_label="BlueBerryPie"
     )
     assert (
-        query
+        str(query)
         == """\
 CREATE CONSTRAINT blue_berry_pie_identifier_uniqueness IF NOT EXISTS
 FOR (n:BlueBerryPie)
@@ -56,7 +64,7 @@ REQUIRE n.identifier IS UNIQUE;"""
 def test_fetch_database_status(query_builder: QueryBuilder) -> None:
     query = query_builder.fetch_database_status()
     assert (
-        query
+        str(query)
         == """\
 SHOW DEFAULT DATABASE
 YIELD currentStatus;"""
@@ -75,7 +83,7 @@ YIELD currentStatus;"""
             True,
             """\
 CALL () {
-    CALL db.index.fulltext.queryNodes("search_index", $query_string)
+    OPTIONAL CALL db.index.fulltext.queryNodes("search_index", $query_string)
     YIELD node AS hit, score
     CALL {
         WITH hit
@@ -116,16 +124,16 @@ CALL () {
         RETURN CASE WHEN referenced IS NOT NULL THEN {
             label: type(r),
             position: r.position,
-            value: referenced.identifier
-        } ELSE NULL END as ref
+            value: referenced_merged_node.identifier
+        } ELSE NULL END AS ref
     UNION
         WITH extracted_node
         OPTIONAL MATCH (extracted_node)-[r]->(nested:Link|Text|Location)
         RETURN CASE WHEN nested IS NOT NULL THEN {
             label: type(r),
             position: r.position,
-            value: properties(nested)
-        } ELSE NULL END as ref
+            value: properties(referenced_nested_node)
+        } ELSE NULL END AS ref
     }
     WITH extracted_node, collect(ref) as refs
     RETURN extracted_node{.*, entityType: head(labels(extracted_node)), _refs: refs}
@@ -151,16 +159,16 @@ CALL {
         RETURN CASE WHEN referenced IS NOT NULL THEN {
             label: type(r),
             position: r.position,
-            value: referenced.identifier
-        } ELSE NULL END as ref
+            value: referenced_merged_node.identifier
+        } ELSE NULL END AS ref
     UNION
         WITH extracted_node
         OPTIONAL MATCH (extracted_node)-[r]->(nested:Link|Text|Location)
         RETURN CASE WHEN nested IS NOT NULL THEN {
             label: type(r),
             position: r.position,
-            value: properties(nested)
-        } ELSE NULL END as ref
+            value: properties(referenced_nested_node)
+        } ELSE NULL END AS ref
     }
     WITH extracted_node, collect(ref) as refs
     RETURN extracted_node{.*, entityType: head(labels(extracted_node)), _refs: refs}
@@ -181,7 +189,7 @@ def test_fetch_extracted_items(
         filter_by_query_string=filter_by_query_string,
         filter_by_stable_target_id=filter_by_stable_target_id,
     )
-    assert query == expected
+    assert str(query) == expected
 
 
 @pytest.mark.parametrize(
@@ -196,92 +204,92 @@ def test_fetch_extracted_items(
             True,
             """\
 CALL () {
-    CALL db.index.fulltext.queryNodes("search_index", $query_string)
+    OPTIONAL CALL db.index.fulltext.queryNodes("search_index", $query_string)
     YIELD node AS hit, score
-    OPTIONAL MATCH (n:AdditiveThis|AdditiveThat|AdditiveOther|ExtractedThis|ExtractedThat|ExtractedOther)-[:stableTargetId]->(merged:MergedThis|MergedThat|MergedOther)
+    OPTIONAL MATCH (extracted_or_rule_node:ExtractedThis|ExtractedThat|ExtractedOther|AdditiveThis|AdditiveThat|AdditiveOther)-[:stableTargetId]->(merged_node:MergedThis|MergedThat|MergedOther)
     WHERE
-        elementId(hit) = elementId(n)
-        AND merged.identifier = $stable_target_id
-        AND ANY(label IN labels(merged) WHERE label IN $labels)
-    WITH DISTINCT merged as merged
-    RETURN COUNT(merged) AS total
+        elementId(hit) = elementId(extracted_or_rule_node)
+        AND merged_node.identifier = $stable_target_id
+        AND ANY(label IN labels(merged_node) WHERE label IN $labels)
+    WITH DISTINCT merged_node AS merged_node
+    RETURN COUNT(merged_node) AS total
 }
 CALL () {
-    CALL db.index.fulltext.queryNodes("search_index", $query_string)
+    OPTIONAL CALL db.index.fulltext.queryNodes("search_index", $query_string)
     YIELD node AS hit, score
-    OPTIONAL MATCH (n:AdditiveThis|AdditiveThat|AdditiveOther|ExtractedThis|ExtractedThat|ExtractedOther)-[:stableTargetId]->(merged:MergedThis|MergedThat|MergedOther)
+    OPTIONAL MATCH (extracted_or_rule_node:ExtractedThis|ExtractedThat|ExtractedOther|AdditiveThis|AdditiveThat|AdditiveOther)-[:stableTargetId]->(merged_node:MergedThis|MergedThat|MergedOther)
     WHERE
-        elementId(hit) = elementId(n)
-        AND merged.identifier = $stable_target_id
-        AND ANY(label IN labels(merged) WHERE label IN $labels)
-    WITH DISTINCT merged as merged
-    OPTIONAL MATCH (n)-[:stableTargetId]->(merged)
-    WITH n, merged
-    CALL (n) {
-        OPTIONAL MATCH (n)-[r]->(referenced:MergedThis|MergedThat|MergedOther)
+        elementId(hit) = elementId(extracted_or_rule_node)
+        AND merged_node.identifier = $stable_target_id
+        AND ANY(label IN labels(merged_node) WHERE label IN $labels)
+    WITH DISTINCT merged_node AS merged_node
+    OPTIONAL MATCH (extracted_or_rule_node)-[:stableTargetId]->(merged_node)
+    WITH extracted_or_rule_node, merged_node
+    CALL (extracted_or_rule_node) {
+        OPTIONAL MATCH (extracted_or_rule_node)-[r]->(referenced:MergedThis|MergedThat|MergedOther)
         RETURN CASE WHEN referenced IS NOT NULL THEN {
             label: type(r),
             position: r.position,
             value: referenced.identifier
-        } ELSE NULL END as ref
+        } ELSE NULL END AS ref
     UNION
-        OPTIONAL MATCH (n)-[r]->(nested:Link|Text|Location)
-        RETURN CASE WHEN nested IS NOT NULL THEN {
+        OPTIONAL MATCH (extracted_or_rule_node)-[r]->(referenced_nested_node:Link|Text|Location)
+        RETURN CASE WHEN referenced_nested_node IS NOT NULL THEN {
             label: type(r),
             position: r.position,
-            value: properties(nested)
-        } ELSE NULL END as ref
+            value: properties(referenced_nested_node)
+        } ELSE NULL END AS ref
     }
-    WITH merged, n, collect(ref) as refs
-    ORDER BY merged.identifier, n.identifier, head(labels(n)) ASC
-    WITH merged, collect(n{.*, entityType: head(labels(n)), _refs: refs}) as n
-    RETURN merged{entityType: head(labels(merged)), identifier: merged.identifier, components: n}
+    WITH merged_node, extracted_or_rule_node, collect(ref) AS refs
+    ORDER BY merged_node.identifier, extracted_or_rule_node.identifier, head(labels(extracted_or_rule_node)) ASC
+    WITH merged_node, collect(extracted_or_rule_node{.*, entityType: head(labels(extracted_or_rule_node)), _refs: refs}) AS extracted_or_rule_node
+    RETURN merged_node{entityType: head(labels(merged_node)), identifier: merged_node.identifier, components: extracted_or_rule_node}
     SKIP $skip
     LIMIT $limit
 }
-RETURN collect(merged) AS items, total;""",
+RETURN collect(merged_node) AS items, total;""",
         ),
         (
             False,
             False,
             """\
 CALL () {
-    OPTIONAL MATCH (n:AdditiveThis|AdditiveThat|AdditiveOther|ExtractedThis|ExtractedThat|ExtractedOther)-[:stableTargetId]->(merged:MergedThis|MergedThat|MergedOther)
+    OPTIONAL MATCH (extracted_or_rule_node:ExtractedThis|ExtractedThat|ExtractedOther|AdditiveThis|AdditiveThat|AdditiveOther)-[:stableTargetId]->(merged_node:MergedThis|MergedThat|MergedOther)
     WHERE
-        ANY(label IN labels(merged) WHERE label IN $labels)
-    WITH DISTINCT merged as merged
-    RETURN COUNT(merged) AS total
+        ANY(label IN labels(merged_node) WHERE label IN $labels)
+    WITH DISTINCT merged_node AS merged_node
+    RETURN COUNT(merged_node) AS total
 }
 CALL () {
-    OPTIONAL MATCH (n:AdditiveThis|AdditiveThat|AdditiveOther|ExtractedThis|ExtractedThat|ExtractedOther)-[:stableTargetId]->(merged:MergedThis|MergedThat|MergedOther)
+    OPTIONAL MATCH (extracted_or_rule_node:ExtractedThis|ExtractedThat|ExtractedOther|AdditiveThis|AdditiveThat|AdditiveOther)-[:stableTargetId]->(merged_node:MergedThis|MergedThat|MergedOther)
     WHERE
-        ANY(label IN labels(merged) WHERE label IN $labels)
-    WITH DISTINCT merged as merged
-    OPTIONAL MATCH (n)-[:stableTargetId]->(merged)
-    WITH n, merged
-    CALL (n) {
-        OPTIONAL MATCH (n)-[r]->(referenced:MergedThis|MergedThat|MergedOther)
+        ANY(label IN labels(merged_node) WHERE label IN $labels)
+    WITH DISTINCT merged_node AS merged_node
+    OPTIONAL MATCH (extracted_or_rule_node)-[:stableTargetId]->(merged_node)
+    WITH extracted_or_rule_node, merged_node
+    CALL (extracted_or_rule_node) {
+        OPTIONAL MATCH (extracted_or_rule_node)-[r]->(referenced:MergedThis|MergedThat|MergedOther)
         RETURN CASE WHEN referenced IS NOT NULL THEN {
             label: type(r),
             position: r.position,
             value: referenced.identifier
-        } ELSE NULL END as ref
+        } ELSE NULL END AS ref
     UNION
-        OPTIONAL MATCH (n)-[r]->(nested:Link|Text|Location)
-        RETURN CASE WHEN nested IS NOT NULL THEN {
+        OPTIONAL MATCH (extracted_or_rule_node)-[r]->(referenced_nested_node:Link|Text|Location)
+        RETURN CASE WHEN referenced_nested_node IS NOT NULL THEN {
             label: type(r),
             position: r.position,
-            value: properties(nested)
-        } ELSE NULL END as ref
+            value: properties(referenced_nested_node)
+        } ELSE NULL END AS ref
     }
-    WITH merged, n, collect(ref) as refs
-    ORDER BY merged.identifier, n.identifier, head(labels(n)) ASC
-    WITH merged, collect(n{.*, entityType: head(labels(n)), _refs: refs}) as n
-    RETURN merged{entityType: head(labels(merged)), identifier: merged.identifier, components: n}
+    WITH merged_node, extracted_or_rule_node, collect(ref) AS refs
+    ORDER BY merged_node.identifier, extracted_or_rule_node.identifier, head(labels(extracted_or_rule_node)) ASC
+    WITH merged_node, collect(extracted_or_rule_node{.*, entityType: head(labels(extracted_or_rule_node)), _refs: refs}) AS extracted_or_rule_node
+    RETURN merged_node{entityType: head(labels(merged_node)), identifier: merged_node.identifier, components: extracted_or_rule_node}
     SKIP $skip
     LIMIT $limit
 }
-RETURN collect(merged) AS items, total;""",
+RETURN collect(merged_node) AS items, total;""",
         ),
     ],
     ids=["all-filters", "no-filters"],
@@ -296,7 +304,7 @@ def test_fetch_merged_items(
         filter_by_query_string=filter_by_query_string,
         filter_by_stable_target_id=filter_by_stable_target_id,
     )
-    assert query == expected
+    assert str(query) == expected
 
 
 @pytest.mark.parametrize(
@@ -319,10 +327,10 @@ WHERE
     AND n.identifierInPrimarySource = $identifier_in_primary_source
     AND merged.identifier = $stable_target_id
 RETURN
-    merged.identifier as stableTargetId,
-    primary_source.identifier as hadPrimarySource,
-    n.identifierInPrimarySource as identifierInPrimarySource,
-    n.identifier as identifier
+    merged.identifier AS stableTargetId,
+    primary_source.identifier AS hadPrimarySource,
+    n.identifierInPrimarySource AS identifierInPrimarySource,
+    n.identifier AS identifier
 ORDER BY n.identifier ASC
 LIMIT $limit;""",
         ),
@@ -334,10 +342,10 @@ LIMIT $limit;""",
 MATCH (n:ExtractedThis|ExtractedThat|ExtractedOther)-[:stableTargetId]->(merged:MergedThis|MergedThat|MergedOther)
 MATCH (n)-[:hadPrimarySource]->(primary_source:MergedPrimarySource)
 RETURN
-    merged.identifier as stableTargetId,
-    primary_source.identifier as hadPrimarySource,
-    n.identifierInPrimarySource as identifierInPrimarySource,
-    n.identifier as identifier
+    merged.identifier AS stableTargetId,
+    primary_source.identifier AS hadPrimarySource,
+    n.identifierInPrimarySource AS identifierInPrimarySource,
+    n.identifier AS identifier
 ORDER BY n.identifier ASC
 LIMIT $limit;""",
         ),
@@ -351,10 +359,10 @@ MATCH (n)-[:hadPrimarySource]->(primary_source:MergedPrimarySource)
 WHERE
     merged.identifier = $stable_target_id
 RETURN
-    merged.identifier as stableTargetId,
-    primary_source.identifier as hadPrimarySource,
-    n.identifierInPrimarySource as identifierInPrimarySource,
-    n.identifier as identifier
+    merged.identifier AS stableTargetId,
+    primary_source.identifier AS hadPrimarySource,
+    n.identifierInPrimarySource AS identifierInPrimarySource,
+    n.identifier AS identifier
 ORDER BY n.identifier ASC
 LIMIT $limit;""",
         ),
@@ -373,7 +381,7 @@ def test_fetch_identities(
         filter_by_identifier_in_primary_source=filter_by_identifier_in_primary_source,
         filter_by_stable_target_id=filter_by_stable_target_id,
     )
-    assert query == expected
+    assert str(query) == expected
 
 
 @pytest.mark.parametrize(
@@ -402,12 +410,12 @@ CALL (source) {
     MERGE (source)-[edge:agendaSignedOff {position: $ref_positions[2]}]->(target_2)
     RETURN edge
 }
-WITH source, count(edge) as merged, collect(edge) as edges
+WITH source, count(edge) AS merged, collect(edge) AS edges
 CALL (source, edges) {
     MATCH (source)-[outdated_edge]->(:MergedThis|MergedThat|MergedOther)
     WHERE NOT outdated_edge IN edges
     DELETE outdated_edge
-    RETURN count(outdated_edge) as pruned
+    RETURN count(outdated_edge) AS pruned
 }
 RETURN merged, pruned, edges;""",
         ),
@@ -416,14 +424,14 @@ RETURN merged, pruned, edges;""",
             """\
 MATCH (source:ExtractedThat {identifier: $identifier})-[:stableTargetId]->({identifier: $stable_target_id})
 CALL (source) {
-    RETURN null as edge
+    RETURN null AS edge
 }
-WITH source, count(edge) as merged, collect(edge) as edges
+WITH source, count(edge) AS merged, collect(edge) AS edges
 CALL (source, edges) {
     MATCH (source)-[outdated_edge]->(:MergedThis|MergedThat|MergedOther)
     WHERE NOT outdated_edge IN edges
     DELETE outdated_edge
-    RETURN count(outdated_edge) as pruned
+    RETURN count(outdated_edge) AS pruned
 }
 RETURN merged, pruned, edges;""",
         ),
@@ -438,7 +446,7 @@ def test_merge_edges(
         current_constraints=["identifier"],
         ref_labels=ref_labels,
     )
-    assert query == expected
+    assert str(query) == expected
 
 
 @pytest.mark.parametrize(
@@ -449,7 +457,7 @@ def test_merge_edges(
             ["Text", "Link", "Location"],
             """\
 MERGE (merged:MergedThat {identifier: $stable_target_id})
-MERGE (current:ExtractedThat {identifier: $identifier})-[stableTargetId:stableTargetId {position: 0}]->(merged)
+MERGE (current:ExtractedThat {identifier: $identifier})-[:stableTargetId {position: 0}]->(merged)
 ON CREATE SET current = $on_create
 ON MATCH SET current += $on_match
 MERGE (current)-[edge_0:description {position: $nested_positions[0]}]->(value_0:Text)
@@ -462,13 +470,13 @@ MERGE (current)-[edge_2:geoLocation {position: $nested_positions[2]}]->(value_2:
 ON CREATE SET value_2 = $nested_values[2]
 ON MATCH SET value_2 += $nested_values[2]
 WITH current,
-    [edge_0, edge_1, edge_2] as edges,
-    [value_0, value_1, value_2] as values
+    [edge_0, edge_1, edge_2] AS edges,
+    [value_0, value_1, value_2] AS values
 CALL (current, values) {
     MATCH (current)-[]->(outdated_node:Link|Text|Location)
     WHERE NOT outdated_node IN values
     DETACH DELETE outdated_node
-    RETURN count(outdated_node) as pruned
+    RETURN count(outdated_node) AS pruned
 }
 RETURN current, edges, values, pruned;""",
         ),
@@ -477,17 +485,17 @@ RETURN current, edges, values, pruned;""",
             [],
             """\
 MERGE (merged:MergedThat {identifier: $stable_target_id})
-MERGE (current:ExtractedThat {identifier: $identifier})-[stableTargetId:stableTargetId {position: 0}]->(merged)
+MERGE (current:ExtractedThat {identifier: $identifier})-[:stableTargetId {position: 0}]->(merged)
 ON CREATE SET current = $on_create
 ON MATCH SET current += $on_match
 WITH current,
-    [] as edges,
-    [] as values
+    [] AS edges,
+    [] AS values
 CALL (current, values) {
     MATCH (current)-[]->(outdated_node:Link|Text|Location)
     WHERE NOT outdated_node IN values
     DETACH DELETE outdated_node
-    RETURN count(outdated_node) as pruned
+    RETURN count(outdated_node) AS pruned
 }
 RETURN current, edges, values, pruned;""",
         ),
@@ -507,4 +515,4 @@ def test_merge_item(
         nested_edge_labels=nested_edge_labels,
         nested_node_labels=nested_node_labels,
     )
-    assert query == expected
+    assert str(query) == expected